refactor(scripts): extract _default_branch to git_utils.py (ticket 0090)
